refactor(daemon): assemble extracted retained owners at the root
