+++ crates/tracedecay/src/daemon/project_open_owners.rs
@@ -399,11 +399,13 @@ pub(super) async fn register_project_open_production_owners(
                 message: format!("project-open retained grant is invalid: {error}"),
             }
         })?;
-    let retained_ports = server.retained_surface_ports(
-        project_root,
-        scope.project_id.clone(),
-        access.configuration_digest.clone(),
-    );
+    let retained_ports = server
+        .retained_surface_ports(
+            project_root,
+            scope.project_id.clone(),
+            access.configuration_digest.clone(),
+        )
+        .await;
     hotpath::future!(
         invocation.retained_runtime_registrar().register(
             project_root.to_path_buf(),
