--- crates/tracedecay/src/daemon/retained_owner/memory.rs
@@ -1,1067 +1,0 @@
-use std::path::PathBuf;
-use std::sync::Arc;
-
-use serde::Serialize;
-use tracedecay_contracts::retained_surfaces::{
-    FactFeedbackRequestV1, FactRetrievalTelemetryV1, FactStoreAddRequestV1,
-    FactStoreContradictRequestV1, FactStoreGetRequestV1, FactStoreListRequestV1,
-    FactStoreProbeRequestV1, FactStoreReasonRequestV1, FactStoreRelatedRequestV1,
-    FactStoreRemoveRequestV1, FactStoreSearchRequestV1, FactStoreSupersedeRequestV1,
-    FactStoreUpdateRequestV1, MemoryScopeV1, MemoryStatusRequestV1, RetainedProjectSelectorV1,
-    RetainedSurfaceOperation, RetainedSurfaceResultV1,
-};
-use tracedecay_contracts::{
-    ApplicationOutcome, RetainedMemoryExecutionPortV1, RetainedMemoryRequestV1,
-    RetainedSurfaceExecutionContextV1, RetainedSurfaceExecutionErrorV1,
-    RetainedSurfaceExecutionFutureV1, now_micros,
-};
-use tracedecay_domain::{FactOwnerV1, ManifestDigest};
-use tracedecay_session_memory::memory::{
-    MemoryApplication, MemoryOperationContext, ProjectMemoryFactAddRequestOutcome,
-};
-use tracedecay_store::{
-    FactReadControl, ProjectMemoryFactContradictionQueryV1,
-    ProjectMemoryFactFeedbackHistoryQueryV1, ProjectMemoryFactIdV1, ProjectMemoryFactListQueryV1,
-    ProjectMemoryFactSearchKindV1,
-};
-
-use super::map_execution_error;
-use super::memory_mapping;
-use super::memory_mutation::{
-    bounded_memory_operation, fact_write_control, validate_memory_mutation,
-};
-use super::memory_tracking::{TrackedExplicitSearch, track_explicit_search};
-use super::receipts::{
-    effective_memory_deadline, evidence_outcome, memory_expiry_partial, prepare_retained_effect,
-};
-use crate::tracedecay::TraceDecay;
-use tracedecay_runtime_core::db::Database;
-use tracedecay_session_memory::fact_store::DatabaseFactStore;
-use tracedecay_store_runtime::DaemonSessionRuntimeRegistryV1;
-
-macro_rules! execute_scoped_memory {
-    (
-        $port:expr,
-        $context:expr,
-        $memory_scope:expr,
-        $selector:expr,
-        $access:expr,
-        $executor:ident($request:expr)
-    ) => {{
-        match &$port.authority {
-            DirectRetainedMemoryAuthorityV1::Project { cg, project_root } => {
-                let (cg, _) = bounded_memory_operation($context, async {
-                    Ok::<_, RetainedSurfaceExecutionErrorV1>(Arc::clone(&*cg.read().await))
-                })
-                .await?;
-                let (target, _) = bounded_memory_operation($context, async {
-                    super::memory_target::open_project_retained_memory_target(
-                        &cg,
-                        project_root,
-                        &$context.request_context.scope().project_id,
-                        $memory_scope,
-                        $selector,
-                        $access,
-                    )
-                    .await
-                })
-                .await?;
-                $executor(
-                    $context,
-                    target.database(),
-                    target.owner().clone(),
-                    $request,
-                    &$port.configuration_digest,
-                )
-                .await
-            }
-            DirectRetainedMemoryAuthorityV1::Profile { registry } => {
-                memory_mapping::ensure_profile_request_scope($memory_scope, $selector)?;
-                let (database, _) = bounded_memory_operation($context, async {
-                    hotpath::future!(
-                        tracedecay_store_runtime::open_user_memory_db(registry),
-                        label = "daemon.retained.memory.open_profile"
-                    )
-                    .await
-                    .map_err(map_execution_error)
-                })
-                .await?;
-                $executor(
-                    $context,
-                    &database,
-                    FactOwnerV1::Profile,
-                    $request,
-                    &$port.configuration_digest,
-                )
-                .await
-            }
-        }
-    }};
-}
-
-enum DirectRetainedMemoryAuthorityV1<'a> {
-    Project {
-        cg: Arc<tokio::sync::RwLock<Arc<TraceDecay>>>,
-        project_root: PathBuf,
-    },
-    Profile {
-        registry: &'a DaemonSessionRuntimeRegistryV1,
-    },
-}
-
-enum Read<'a> {
-    Search(&'a FactStoreSearchRequestV1),
-    Probe(&'a FactStoreProbeRequestV1),
-    Related(&'a FactStoreRelatedRequestV1),
-    Reason(&'a FactStoreReasonRequestV1),
-    Contradict(&'a FactStoreContradictRequestV1),
-    Get(&'a FactStoreGetRequestV1),
-    List(&'a FactStoreListRequestV1),
-}
-#[derive(Clone, Copy)]
-enum SemanticRead<'a> {
-    Probe(&'a FactStoreProbeRequestV1),
-    Related(&'a FactStoreRelatedRequestV1),
-    Reason(&'a FactStoreReasonRequestV1),
-}
-impl Read<'_> {
-    fn scope(&self) -> (Option<MemoryScopeV1>, Option<&RetainedProjectSelectorV1>) {
-        match self {
-            Self::Search(request) => memory_mapping::read_scope(&request.options),
-            Self::Probe(request) => memory_mapping::read_scope(&request.options),
-            Self::Related(request) => memory_mapping::read_scope(&request.options),
-            Self::Reason(request) => memory_mapping::read_scope(&request.options),
-            Self::Contradict(request) => (request.memory_scope, request.project_selector.as_ref()),
-            Self::Get(request) => (request.memory_scope, request.project_selector.as_ref()),
-            Self::List(request) => memory_mapping::read_scope(&request.options),
-        }
-    }
-}
-pub(super) struct DirectRetainedMemoryPortV1<'a> {
-    authority: DirectRetainedMemoryAuthorityV1<'a>,
-    configuration_digest: ManifestDigest,
-}
-impl DirectRetainedMemoryPortV1<'static> {
-    pub(super) fn project(
-        cg: Arc<tokio::sync::RwLock<Arc<TraceDecay>>>,
-        project_root: PathBuf,
-        configuration_digest: ManifestDigest,
-    ) -> Self {
-        Self {
-            authority: DirectRetainedMemoryAuthorityV1::Project { cg, project_root },
-            configuration_digest,
-        }
-    }
-}
-
-impl<'a> DirectRetainedMemoryPortV1<'a> {
-    pub(super) fn profile(
-        registry: &'a DaemonSessionRuntimeRegistryV1,
-        configuration_digest: ManifestDigest,
-    ) -> Self {
-        Self {
-            authority: DirectRetainedMemoryAuthorityV1::Profile { registry },
-            configuration_digest,
-        }
-    }
-
-    #[hotpath::skip]
-    async fn execute_add(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &FactStoreAddRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Write,
-            execute_add_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_read(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: Read<'_>,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        let (memory_scope, selector) = request.scope();
-        execute_scoped_memory!(
-            self,
-            context,
-            memory_scope,
-            selector,
-            super::memory_target::MemoryTargetAccessV1::Read,
-            execute_read_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_status(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &MemoryStatusRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Read,
-            execute_status_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_update(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &FactStoreUpdateRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Write,
-            execute_update_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_remove(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &FactStoreRemoveRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Write,
-            execute_remove_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_supersede(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &FactStoreSupersedeRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Write,
-            execute_supersede_on_db(request)
-        )
-    }
-
-    #[hotpath::skip]
-    async fn execute_feedback(
-        &self,
-        context: &RetainedSurfaceExecutionContextV1<'_>,
-        request: &FactFeedbackRequestV1,
-    ) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-        execute_scoped_memory!(
-            self,
-            context,
-            request.memory_scope,
-            request.project_selector.as_ref(),
-            super::memory_target::MemoryTargetAccessV1::Write,
-            execute_feedback_on_db(request)
-        )
-    }
-}
-
-impl RetainedMemoryExecutionPortV1 for DirectRetainedMemoryPortV1<'_> {
-    fn execute_memory<'a>(
-        &'a self,
-        context: RetainedSurfaceExecutionContextV1<'a>,
-        request: RetainedMemoryRequestV1<'a>,
-    ) -> RetainedSurfaceExecutionFutureV1<'a> {
-        Box::pin(async move {
-            match request {
-                RetainedMemoryRequestV1::FactStoreAdd(request) => {
-                    self.execute_add(&context, request).await
-                }
-                RetainedMemoryRequestV1::FactStoreSearch(request) => {
-                    self.execute_read(&context, Read::Search(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreProbe(request) => {
-                    self.execute_read(&context, Read::Probe(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreRelated(request) => {
-                    self.execute_read(&context, Read::Related(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreReason(request) => {
-                    self.execute_read(&context, Read::Reason(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreContradict(request) => {
-                    self.execute_read(&context, Read::Contradict(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreGet(request) => {
-                    self.execute_read(&context, Read::Get(request)).await
-                }
-                RetainedMemoryRequestV1::FactStoreUpdate(request) => {
-                    self.execute_update(&context, request).await
-                }
-                RetainedMemoryRequestV1::FactStoreRemove(request) => {
-                    self.execute_remove(&context, request).await
-                }
-                RetainedMemoryRequestV1::FactStoreSupersede(request) => {
-                    self.execute_supersede(&context, request).await
-                }
-                RetainedMemoryRequestV1::FactStoreList(request) => {
-                    self.execute_read(&context, Read::List(request)).await
-                }
-                RetainedMemoryRequestV1::FactFeedback(request) => {
-                    self.execute_feedback(&context, request).await
-                }
-                RetainedMemoryRequestV1::MemoryStatus(request) => {
-                    self.execute_status(&context, request).await
-                }
-            }
-        })
-    }
-}
-
-async fn execute_add_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreAddRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let add_request = memory_mapping::add_request(request)?;
-    let preflight = memory
-        .preflight_project_memory_fact_add(
-            add_request,
-            Some(context.request_context.actor().clone()),
-        )
-        .map_err(memory_mapping::map_memory_error)?;
-    let operation_id = preflight.operation_id().as_str().to_owned();
-    let prepared = prepare_retained_effect(
-        context,
-        RetainedSurfaceOperation::FactStoreAdd,
-        configuration_digest,
-        preflight.effect_material(),
-        &operation_id,
-    )?;
-    let write_control = fact_write_control(context);
-    let (outcome, settled_after_expiry) = bounded_memory_operation(context, async {
-        Ok(hotpath::future!(
-            memory.add_preflighted_project_memory_fact(preflight, &write_control),
-            label = "daemon.retained.memory.add.commit"
-        )
-        .await)
-    })
-    .await?;
-    let outcome = validate_memory_mutation(outcome, &prepared, |outcome| match outcome {
-        ProjectMemoryFactAddRequestOutcome::RejectedSecretLike => None,
-        ProjectMemoryFactAddRequestOutcome::Applied(outcome) => outcome
-            .commit_receipt()
-            .map(tracedecay_store::FactCommitReceipt::committed_state_digest),
-    })?;
-    let committed_receipt = match &outcome {
-        ProjectMemoryFactAddRequestOutcome::RejectedSecretLike => None,
-        ProjectMemoryFactAddRequestOutcome::Applied(outcome) => outcome.commit_receipt(),
-    };
-    let public = match memory_mapping::add_result(&outcome) {
-        Ok(public) => public,
-        Err(error) => {
-            let Some(commit) = committed_receipt else {
-                return Err(error);
-            };
-            return prepared.memory_projection_failed(commit.committed_state_digest());
-        }
-    };
-    let result = RetainedSurfaceResultV1::FactStoreAdd(public.clone());
-    let committed_state = match committed_receipt {
-        Some(commit) => commit.committed_state_digest().clone(),
-        None => prepared
-            .material_committed_state_digest(&memory_mapping::add_committed_state(&outcome)?)?,
-    };
-    prepared.complete_with_digest(
-        context,
-        &committed_state,
-        tracedecay_contracts::ReconciliationState::Reconciled,
-        result,
-        memory_expiry_partial(settled_after_expiry),
-    )
-}
-
-async fn execute_update_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreUpdateRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let update = memory_mapping::PreparedFactUpdate::new(owner, request)?;
-    let logical_effect = update.logical_effect()?;
-    let operation_context =
-        memory_operation_context(context, update.owner(), "update", &logical_effect)?;
-    let operation_id = operation_context.operation_id().as_str().to_owned();
-    let prepared = prepare_retained_effect(
-        context,
-        RetainedSurfaceOperation::FactStoreUpdate,
-        configuration_digest,
-        &logical_effect,
-        &operation_id,
-    )?;
-    let command = update.into_command(
-        operation_context.operation_id().clone(),
-        context.request_context.actor().clone(),
-    )?;
-    let write_control = fact_write_control(context);
-    let (outcome, settled_after_expiry) = bounded_memory_operation(context, async {
-        Ok(hotpath::future!(
-            memory.update_project_memory_fact(command, &write_control),
-            label = "daemon.retained.memory.update.commit"
-        )
-        .await)
-    })
-    .await?;
-    let outcome = validate_memory_mutation(outcome, &prepared, |outcome| {
-        Some(outcome.commit_receipt().committed_state_digest())
-    })?;
-    let commit = outcome.commit_receipt();
-    let result = match memory_mapping::update_result(&outcome) {
-        Ok(result) => RetainedSurfaceResultV1::FactStoreUpdate(result),
-        Err(_) => {
-            return prepared.memory_projection_failed(commit.committed_state_digest());
-        }
-    };
-    prepared.complete_with_digest(
-        context,
-        commit.committed_state_digest(),
-        tracedecay_contracts::ReconciliationState::Reconciled,
-        result,
-        memory_expiry_partial(settled_after_expiry),
-    )
-}
-
-async fn execute_remove_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreRemoveRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let remove = memory_mapping::PreparedFactRemove::new(owner, request)?;
-    let logical_effect = remove.logical_effect()?;
-    let operation_context =
-        memory_operation_context(context, remove.owner(), "remove", &logical_effect)?;
-    let operation_id = operation_context.operation_id().as_str().to_owned();
-    let prepared = prepare_retained_effect(
-        context,
-        RetainedSurfaceOperation::FactStoreRemove,
-        configuration_digest,
-        &logical_effect,
-        &operation_id,
-    )?;
-    let command = remove.into_command(
-        operation_context.operation_id().clone(),
-        context.request_context.actor().clone(),
-    )?;
-    let write_control = fact_write_control(context);
-    let (outcome, settled_after_expiry) = bounded_memory_operation(context, async {
-        Ok(hotpath::future!(
-            memory.remove_project_memory_fact(command, &write_control),
-            label = "daemon.retained.memory.remove.commit"
-        )
-        .await)
-    })
-    .await?;
-    let outcome = validate_memory_mutation(outcome, &prepared, |outcome| {
-        outcome
-            .commit_receipt()
-            .map(tracedecay_store::FactCommitReceipt::committed_state_digest)
-    })?;
-    let committed_receipt = outcome.commit_receipt();
-    let public = match memory_mapping::remove_result(&outcome) {
-        Ok(public) => public,
-        Err(error) => {
-            let Some(commit) = committed_receipt else {
-                return Err(error);
-            };
-            return prepared.memory_projection_failed(commit.committed_state_digest());
-        }
-    };
-    let result = RetainedSurfaceResultV1::FactStoreRemove(public.clone());
-    let partial = memory_expiry_partial(settled_after_expiry);
-    if let Some(commit) = committed_receipt {
-        return prepared.complete_with_digest(
-            context,
-            commit.committed_state_digest(),
-            tracedecay_contracts::ReconciliationState::Reconciled,
-            result,
-            partial,
-        );
-    }
-    prepared.complete(
-        context,
-        &public,
-        tracedecay_contracts::ReconciliationState::Reconciled,
-        result,
-        partial,
-    )
-}
-
-async fn execute_supersede_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreSupersedeRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let supersede = memory_mapping::PreparedFactSupersede::new(owner, request)?;
-    let logical_effect = supersede.logical_effect()?;
-    let operation_context =
-        memory_operation_context(context, supersede.owner(), "supersede", &logical_effect)?;
-    let operation_id = operation_context.operation_id().as_str().to_owned();
-    let prepared = prepare_retained_effect(
-        context,
-        RetainedSurfaceOperation::FactStoreSupersede,
-        configuration_digest,
-        &logical_effect,
-        &operation_id,
-    )?;
-    let command = supersede.into_command(
-        operation_context.operation_id().clone(),
-        context.request_context.actor().clone(),
-    )?;
-    let write_control = fact_write_control(context);
-    let (outcome, settled_after_expiry) = bounded_memory_operation(context, async {
-        Ok(hotpath::future!(
-            memory.supersede_project_memory_fact(command, &write_control),
-            label = "daemon.retained.memory.supersede.commit"
-        )
-        .await)
-    })
-    .await?;
-    let outcome = validate_memory_mutation(outcome, &prepared, |outcome| {
-        outcome
-            .commit_receipt()
-            .map(tracedecay_store::FactCommitReceipt::committed_state_digest)
-    })?;
-    let committed_receipt = outcome.commit_receipt();
-    let public = memory_mapping::supersede_result(&outcome)?;
-    let result = RetainedSurfaceResultV1::FactStoreSupersede(public.clone());
-    let partial = memory_expiry_partial(settled_after_expiry);
-    if let Some(commit) = committed_receipt {
-        return prepared.complete_with_digest(
-            context,
-            commit.committed_state_digest(),
-            tracedecay_contracts::ReconciliationState::Reconciled,
-            result,
-            partial,
-        );
-    }
-    prepared.complete(
-        context,
-        &public,
-        tracedecay_contracts::ReconciliationState::Reconciled,
-        result,
-        partial,
-    )
-}
-
-async fn execute_feedback_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactFeedbackRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let feedback = memory_mapping::PreparedFactFeedback::new(owner, request)?;
-    let logical_effect = feedback.logical_effect()?;
-    let operation_context =
-        memory_operation_context(context, feedback.owner(), "feedback", &logical_effect)?;
-    let operation_id = operation_context.operation_id().as_str().to_owned();
-    let prepared = prepare_retained_effect(
-        context,
-        RetainedSurfaceOperation::FactFeedback,
-        configuration_digest,
-        &logical_effect,
-        &operation_id,
-    )?;
-    let command = feedback.into_command(
-        operation_context.operation_id().clone(),
-        context.request_context.actor().clone(),
-    )?;
-    let write_control = fact_write_control(context);
-    let (outcome, settled_after_expiry) = bounded_memory_operation(context, async {
-        Ok(hotpath::future!(
-            memory.record_project_memory_fact_feedback(command, &write_control),
-            label = "daemon.retained.memory.feedback.commit"
-        )
-        .await)
-    })
-    .await?;
-    let outcome = validate_memory_mutation(outcome, &prepared, |outcome| {
-        Some(outcome.commit_receipt().committed_state_digest())
-    })?;
-    let commit = outcome.commit_receipt();
-    let result = match memory_mapping::feedback_result(&outcome, request.action) {
-        Ok(result) => RetainedSurfaceResultV1::FactFeedback(result),
-        Err(_) => {
-            return prepared.memory_projection_failed(commit.committed_state_digest());
-        }
-    };
-    prepared.complete_with_digest(
-        context,
-        commit.committed_state_digest(),
-        tracedecay_contracts::ReconciliationState::Reconciled,
-        result,
-        memory_expiry_partial(settled_after_expiry),
-    )
-}
-
-async fn execute_read_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: Read<'_>,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    match request {
-        Read::Search(request) => {
-            search_on_db(context, database, owner, request, configuration_digest).await
-        }
-        Read::Probe(request) => {
-            semantic_search_on_db(context, database, owner, SemanticRead::Probe(request)).await
-        }
-        Read::Related(request) => {
-            semantic_search_on_db(context, database, owner, SemanticRead::Related(request)).await
-        }
-        Read::Reason(request) => {
-            semantic_search_on_db(context, database, owner, SemanticRead::Reason(request)).await
-        }
-        Read::Contradict(request) => contradict_on_db(context, database, owner, request).await,
-        Read::Get(request) => get_on_db(context, database, owner, request).await,
-        Read::List(request) => list_on_db(context, database, owner, request).await,
-    }
-}
-
-async fn search_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreSearchRequestV1,
-    configuration_digest: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let search = memory_mapping::PreparedFactSearch::new(owner.clone(), request)?;
-    let logical_effect = search.logical_effect()?;
-    let query = search.into_query();
-    let request_id = context.request_context.request_id().as_str();
-    let actor = Some(context.request_context.actor().clone());
-    let operation_context =
-        MemoryOperationContext::from_request_id(&owner, "search", request_id, actor)
-            .map_err(memory_mapping::map_memory_error)?;
-    let read_control = fact_read_control(context);
-    let (page, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.search_project_memory_facts(query, &read_control),
-            label = "daemon.retained.memory.search.query"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let prepared = if page.hits().is_empty() {
-        None
-    } else {
-        Some(prepare_retained_effect(
-            context,
-            RetainedSurfaceOperation::FactStoreSearch,
-            configuration_digest,
-            &logical_effect,
-            operation_context.operation_id().as_str(),
-        )?)
-    };
-    // Retrieval telemetry is recall bookkeeping for the returned hits, not
-    // part of the evidence itself. When only its write lane is unavailable,
-    // the search degrades to delivering the evidence with a typed telemetry
-    // state instead of refusing a read the store already served.
-    let (tracked, retrieval_telemetry) = if page.hits().is_empty() {
-        (
-            TrackedExplicitSearch::default(),
-            FactRetrievalTelemetryV1::NotApplicable,
-        )
-    } else if database.is_writable() {
-        match track_explicit_search(
-            context,
-            &memory,
-            &owner,
-            operation_context.operation_id().clone(),
-            &page,
-        )
-        .await
-        {
-            Ok(tracked) => {
-                let fact_count = page.hits().len();
-                (tracked, FactRetrievalTelemetryV1::Recorded { fact_count })
-            }
-            Err(error) => match memory_mapping::retrieval_telemetry_degradation(&error) {
-                Some(reason) => (
-                    TrackedExplicitSearch::default(),
-                    FactRetrievalTelemetryV1::Degraded { reason },
-                ),
-                None => return Err(error),
-            },
-        }
-    } else {
-        (
-            TrackedExplicitSearch::default(),
-            FactRetrievalTelemetryV1::ReadOnly,
-        )
-    };
-    if tracked.authority_result_invalid {
-        let committed_state = tracked.committed_state().ok_or_else(|| {
-            RetainedSurfaceExecutionErrorV1::unavailable(
-                "the tracked retrieval telemetry carried no committed state",
-            )
-        })?;
-        let prepared = prepared.as_ref().ok_or_else(|| {
-            RetainedSurfaceExecutionErrorV1::unavailable(
-                "the retained memory effect was not prepared before settlement",
-            )
-        })?;
-        return prepared.partial_with_digest(
-            committed_state,
-            "application.retained.memory-search-authority-result-invalid",
-            "Retrieval telemetry committed, but the authority result failed validation.",
-        );
-    }
-    let mut mapped = match memory_mapping::search_page(&page) {
-        Ok(mapped) => mapped,
-        Err(error) => {
-            let Some(committed_state) = tracked.committed_state() else {
-                return Err(error);
-            };
-            let prepared = prepared.as_ref().ok_or_else(|| {
-                RetainedSurfaceExecutionErrorV1::unavailable(
-                    "the retained memory effect was not prepared before settlement",
-                )
-            })?;
-            return prepared.partial_with_digest(
-                committed_state,
-                "application.retained.memory-search-projection-failed",
-                "Retrieval telemetry committed, but the public search projection could not be assembled.",
-            );
-        }
-    };
-    if tracked.receipt.is_some()
-        && memory_mapping::refresh_search_hits(&mut mapped, &tracked.projections).is_err()
-    {
-        let Some(committed_state) = tracked.committed_state() else {
-            return Err(RetainedSurfaceExecutionErrorV1::unavailable(
-                "the refreshed search hits could not be projected and no telemetry state committed",
-            ));
-        };
-        let prepared = prepared.as_ref().ok_or_else(|| {
-            RetainedSurfaceExecutionErrorV1::unavailable(
-                "the retained memory effect was not prepared before settlement",
-            )
-        })?;
-        return prepared.partial_with_digest(
-            committed_state,
-            "application.retained.memory-search-telemetry-projection-failed",
-            "Retrieval telemetry committed, but the refreshed public facts could not be assembled.",
-        );
-    }
-    if tracked.settled_after_expiry {
-        let Some(committed_state) = tracked.committed_state() else {
-            return Err(RetainedSurfaceExecutionErrorV1::TimedOut(
-                tracedecay_contracts::CancellationStage::DuringRead,
-            ));
-        };
-        let prepared = prepared.as_ref().ok_or_else(|| {
-            RetainedSurfaceExecutionErrorV1::unavailable(
-                "the retained memory effect was not prepared before settlement",
-            )
-        })?;
-        return prepared.partial_with_digest(
-            committed_state,
-            "application.retained.memory-search-admission-expiry-after-telemetry-commit",
-            "Retrieval telemetry committed after the request or capability grant expired.",
-        );
-    }
-    let result = memory_mapping::exact_search_result(mapped, retrieval_telemetry);
-    match evidence_outcome(context, RetainedSurfaceOperation::FactStoreSearch, result) {
-        Ok(outcome) => Ok(outcome),
-        Err(RetainedSurfaceExecutionErrorV1::TimedOut(
-            tracedecay_contracts::CancellationStage::DuringRead,
-        )) => {
-            let Some(committed_state) = tracked.committed_state() else {
-                return Err(RetainedSurfaceExecutionErrorV1::TimedOut(
-                    tracedecay_contracts::CancellationStage::DuringRead,
-                ));
-            };
-            let prepared = prepared.as_ref().ok_or_else(|| {
-                RetainedSurfaceExecutionErrorV1::unavailable(
-                    "the retained memory effect was not prepared before settlement",
-                )
-            })?;
-            prepared.memory_expiry_failed(committed_state)
-        }
-        Err(error) => {
-            let Some(committed_state) = tracked.committed_state() else {
-                return Err(error);
-            };
-            let prepared = prepared.as_ref().ok_or_else(|| {
-                RetainedSurfaceExecutionErrorV1::unavailable(
-                    "the retained memory effect was not prepared before settlement",
-                )
-            })?;
-            prepared.partial_with_digest(
-                committed_state,
-                "application.retained.memory-search-delivery-failed",
-                "Retrieval telemetry committed, but the evidence packet could not be assembled.",
-            )
-        }
-    }
-}
-
-async fn semantic_search_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: SemanticRead<'_>,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let (kind, query_text, options, after, operation) = match request {
-        SemanticRead::Probe(request) => (
-            ProjectMemoryFactSearchKindV1::Probe,
-            Some(request.entity.clone()),
-            &request.options,
-            request.after.as_ref(),
-            RetainedSurfaceOperation::FactStoreProbe,
-        ),
-        SemanticRead::Related(request) => (
-            ProjectMemoryFactSearchKindV1::Related {
-                entity: request.entity.clone(),
-            },
-            None,
-            &request.options,
-            request.after.as_ref(),
-            RetainedSurfaceOperation::FactStoreRelated,
-        ),
-        SemanticRead::Reason(request) => {
-            memory_mapping::validate_reason_entities(&request.entities)?;
-            let entities = request.entities.clone();
-            (
-                ProjectMemoryFactSearchKindV1::Reason { entities },
-                None,
-                &request.options,
-                request.after.as_ref(),
-                RetainedSurfaceOperation::FactStoreReason,
-            )
-        }
-    };
-    let memory = memory_application(database, owner.clone())?;
-    let query = memory_mapping::search_query(owner, kind, query_text, options, after)?;
-    let read_control = fact_read_control(context);
-    let (page, _) = bounded_memory_operation(context, async {
-        let page = match request {
-            SemanticRead::Probe(_) => {
-                hotpath::future!(
-                    memory.probe_project_memory_facts(query, &read_control),
-                    label = "daemon.retained.memory.probe"
-                )
-                .await
-            }
-            SemanticRead::Related(_) => {
-                hotpath::future!(
-                    memory.related_project_memory_facts(query, &read_control),
-                    label = "daemon.retained.memory.related"
-                )
-                .await
-            }
-            SemanticRead::Reason(_) => {
-                hotpath::future!(
-                    memory.reason_project_memory_facts(query, &read_control),
-                    label = "daemon.retained.memory.reason"
-                )
-                .await
-            }
-        };
-        page.map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let mapped = memory_mapping::search_page(&page)?;
-    let result = memory_mapping::semantic_search_result(operation, mapped)?;
-    evidence_outcome(context, operation, result)
-}
-
-async fn contradict_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreContradictRequestV1,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let query = ProjectMemoryFactContradictionQueryV1::new(
-        owner,
-        request.category,
-        request.threshold_millionths.unwrap_or(300_000),
-        memory_mapping::fact_limit(request.limit)?,
-    )
-    .map_err(memory_mapping::map_store_error)?;
-    let read_control = fact_read_control(context);
-    let (page, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.find_project_memory_contradictions(query, &read_control),
-            label = "daemon.retained.memory.contradict.query"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let result =
-        RetainedSurfaceResultV1::FactStoreContradict(memory_mapping::contradiction_page(&page)?);
-    evidence_outcome(
-        context,
-        RetainedSurfaceOperation::FactStoreContradict,
-        result,
-    )
-}
-
-async fn get_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreGetRequestV1,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let target = ProjectMemoryFactIdV1::new(owner, request.fact_id.clone())
-        .map_err(memory_mapping::map_store_error)?;
-    let read_control = fact_read_control(context);
-    let (projection, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.get_project_memory_fact(target.clone(), &read_control),
-            label = "daemon.retained.memory.get.fact"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let projection = projection.ok_or(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized)?;
-    let history_query = ProjectMemoryFactFeedbackHistoryQueryV1::new(
-        target,
-        None,
-        memory_mapping::MAX_RETAINED_FEEDBACK_HISTORY_LIMIT,
-    )
-    .map_err(memory_mapping::map_store_error)?;
-    let (history, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.get_project_memory_feedback_history(history_query, &read_control),
-            label = "daemon.retained.memory.get.history"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let result = memory_mapping::get_result(&projection, &history)?;
-    evidence_outcome(context, RetainedSurfaceOperation::FactStoreGet, result)
-}
-
-async fn list_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    request: &FactStoreListRequestV1,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner.clone())?;
-    let query = ProjectMemoryFactListQueryV1::new(
-        owner,
-        request.options.category,
-        memory_mapping::confidence(request.options.min_trust)?,
-        request.after_fact_id.clone(),
-        memory_mapping::fact_limit(request.options.limit)?,
-    )
-    .map_err(memory_mapping::map_store_error)?;
-    let read_control = fact_read_control(context);
-    let (page, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.list_project_memory_facts(query, &read_control),
-            label = "daemon.retained.memory.list.query"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let result = RetainedSurfaceResultV1::FactStoreList(memory_mapping::list_page(&page)?);
-    evidence_outcome(context, RetainedSurfaceOperation::FactStoreList, result)
-}
-
-async fn execute_status_on_db(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    database: &Database,
-    owner: FactOwnerV1,
-    _: &MemoryStatusRequestV1,
-    _: &ManifestDigest,
-) -> Result<ApplicationOutcome<RetainedSurfaceResultV1>, RetainedSurfaceExecutionErrorV1> {
-    let memory = memory_application(database, owner)?;
-    let read_control = fact_read_control(context);
-    let (status, _) = bounded_memory_operation(context, async {
-        hotpath::future!(
-            memory.project_memory_status(&read_control),
-            label = "daemon.retained.memory.status.query"
-        )
-        .await
-        .map_err(memory_mapping::map_memory_error)
-    })
-    .await?;
-    let result = memory_mapping::status_result(&status);
-    evidence_outcome(context, RetainedSurfaceOperation::MemoryStatus, result)
-}
-
-fn memory_application(
-    database: &Database,
-    owner: FactOwnerV1,
-) -> Result<MemoryApplication<DatabaseFactStore<'_>>, RetainedSurfaceExecutionErrorV1> {
-    MemoryApplication::new(owner, DatabaseFactStore::new(database))
-        .map_err(memory_mapping::map_memory_error)
-}
-
-fn fact_read_control(context: &RetainedSurfaceExecutionContextV1<'_>) -> FactReadControl {
-    let signal = context.cancellation_signal.clone();
-    let expires_at = effective_expiry(context);
-    FactReadControl::new(Arc::new(move || {
-        signal.is_cancelled() || expires_at <= now_micros()
-    }))
-}
-
-fn effective_expiry(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-) -> tracedecay_domain::UtcMicros {
-    effective_memory_deadline(context).expires_at
-}
-
-fn memory_operation_context<T: Serialize>(
-    context: &RetainedSurfaceExecutionContextV1<'_>,
-    owner: &FactOwnerV1,
-    action: &str,
-    logical_effect: &T,
-) -> Result<MemoryOperationContext, RetainedSurfaceExecutionErrorV1> {
-    MemoryOperationContext::from_logical_effect(
-        owner,
-        action,
-        logical_effect,
-        Some(context.request_context.actor().clone()),
-    )
-    .map_err(memory_mapping::map_memory_error)
-}
