--- crates/tracedecay/src/daemon/retained_owner/session_refresh.rs
@@ -1,337 +1,0 @@
-//! Daemon-owned admission boundary for retained session refresh operations.
-//!
-//! The mounted service is shared with MCP, while this adapter independently
-//! binds the retained request to the canonical session application context of
-//! the project- or profile-owned session store that serves it.
-
-use std::collections::BTreeSet;
-
-use sha2::{Digest, Sha256};
-use tracedecay_contracts::retained_surfaces::{
-    RetainedSurfaceExecutionErrorV1, SessionRefreshActionRequestV1, SessionRefreshActionV1,
-    SessionRefreshGrainV1, SessionRefreshRequestV1, SessionRefreshScopeV1,
-    SessionRefreshTemporalModeV1,
-};
-use tracedecay_contracts::{
-    CancellationContext, CancellationSignal, CapabilityGrantId, CapabilityGrantSnapshot, Deadline,
-    DisclosureClass, RequestContext, retained_surface_application_operation,
-};
-use tracedecay_domain::{
-    ManifestDigest, ProjectId, RepositoryId, RetrievalGrainV1, SessionId, TemporalModeV1,
-    UserProfileId, UtcMicros, WorktreeId,
-};
-use tracedecay_session_memory::context::{
-    BranchId, CancellationToken, CapabilityDigest, ConfigurationDigest, PolicyDigest, ProfileId,
-    RequestBudgets, ResolvedGitRoute, ResolvedSessionIdentity, SessionRootId, SessionStoreId,
-    session_application_grant_digest,
-};
-use tracedecay_session_memory::session::{SessionRefreshTarget, SessionRequestBinding};
-use tracedecay_store::SessionRefreshFrontierV1;
-
-pub(crate) use tracedecay_session_memory::session::SessionRefreshServicePort as RetainedSessionRefreshPortV1;
-use tracedecay_session_memory::session::{SessionRefreshAction, SessionRefreshCommand};
-
-const REQUEST_MAX_RESULTS: u64 = 64;
-const REQUEST_MAX_BYTES: u64 = 64 * 1024 * 1024;
-const REQUEST_MAX_WORK_UNITS: u64 = 10_000;
-// One opaque handle crosses the action-specific begin, status, and cancel operations.
-const SESSION_REFRESH_LIFECYCLE_CAPABILITY: &[u8] =
-    b"application.retained.session-refresh-lifecycle.v1";
-
-/// The exact mounted owner a refresh request must name before it is admitted.
-///
-/// `policy_digest` is the stable admission authority the opaque handle is
-/// bound to: the project-open capability grant for a project owner, and the
-/// connection configuration digest for a profile owner, whose per-request
-/// grants would otherwise rebind every `status`/`cancel` to a fresh join key.
-pub(super) struct MountedSessionRefreshAuthorityV1<'a> {
-    pub(super) profile_id: &'a UserProfileId,
-    pub(super) session_store_id: &'a SessionStoreId,
-    pub(super) session_root_id: &'a SessionRootId,
-    pub(super) configuration_digest: &'a ManifestDigest,
-    pub(super) policy_digest: &'a ManifestDigest,
-    pub(super) refresh: &'a dyn RetainedSessionRefreshPortV1,
-}
-
-#[hotpath::measure(label = "daemon.retained.session.refresh_admit")]
-pub(super) fn admitted_session_refresh_command(
-    request: &SessionRefreshRequestV1,
-    context: &RequestContext,
-    cancellation_signal: &CancellationSignal,
-    mounted: &MountedSessionRefreshAuthorityV1<'_>,
-) -> Result<SessionRefreshCommand, RetainedSurfaceExecutionErrorV1> {
-    let admitted = session_refresh_command(request, context, cancellation_signal, mounted);
-    // Refused admissions are recorded so trigger volume that never reaches
-    // refresh execution stays visible in profiles.
-    match &admitted {
-        Ok(_) => {}
-        Err(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized) => {
-            hotpath::gauge!("daemon.retained.session.refresh_admit.denied").inc(1.0);
-        }
-        Err(_) => {
-            hotpath::gauge!("daemon.retained.session.refresh_admit.refused").inc(1.0);
-        }
-    }
-    admitted
-}
-
-fn session_refresh_command(
-    request: &SessionRefreshRequestV1,
-    context: &RequestContext,
-    cancellation_signal: &CancellationSignal,
-    mounted: &MountedSessionRefreshAuthorityV1<'_>,
-) -> Result<SessionRefreshCommand, RetainedSurfaceExecutionErrorV1> {
-    if cancellation_signal.context().token_id != context.cancellation().token_id {
-        return Err(RetainedSurfaceExecutionErrorV1::InvalidRequest);
-    }
-    if request.request.scope.profile_id() != mounted.profile_id.as_str() {
-        return Err(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized);
-    }
-    if request.request.session.store_id != mounted.session_store_id.as_str()
-        || request.request.session.root_id != mounted.session_root_id.as_str()
-    {
-        return Err(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized);
-    }
-    if !request_matches_mounted_project_scope(request, context) {
-        return Err(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized);
-    }
-
-    let selectors = &request.request;
-    let action = admitted_action(request.action, selectors.handle.as_deref())?;
-
-    // A profile-owned identity resolves to the synthetic profile-session scope,
-    // so this equality is what binds a profile refresh to the profile
-    // authority's own admitted context rather than to any project.
-    let identity = admitted_identity(selectors)?;
-    let resolved_scope = identity
-        .session_request_scope()
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    if resolved_scope != *context.scope() {
-        return Err(RetainedSurfaceExecutionErrorV1::NotFoundOrNotAuthorized);
-    }
-    let target = admitted_target(selectors)?;
-    let operation = retained_surface_application_operation(request.operation())
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::Unsupported)?;
-    let capability_digest = CapabilityDigest::new(admitted_digest(
-        b"tracedecay.retained.session-refresh.capability.v1\0",
-        SESSION_REFRESH_LIFECYCLE_CAPABILITY,
-    ));
-    let policy_digest = PolicyDigest::new(admitted_digest(
-        b"tracedecay.retained.session-refresh.policy.v1\0",
-        mounted.policy_digest.as_str().as_bytes(),
-    ));
-    let configuration_digest = ConfigurationDigest::new(admitted_digest(
-        b"tracedecay.retained.session-refresh.configuration.v1\0",
-        mounted.configuration_digest.as_str().as_bytes(),
-    ));
-    let cancellation = CancellationToken::for_application_request(context.request_id().as_str());
-    if context.cancellation().is_cancelled() {
-        cancellation.cancel();
-    }
-    let budgets = RequestBudgets::new(
-        REQUEST_MAX_RESULTS,
-        REQUEST_MAX_BYTES,
-        REQUEST_MAX_WORK_UNITS,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let grant_digest = session_application_grant_digest(
-        capability_digest,
-        policy_digest,
-        configuration_digest,
-        &cancellation,
-        budgets,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let expires_at = std::cmp::min(context.grant().expires_at, context.deadline().expires_at);
-    let grant = CapabilityGrantSnapshot::new(
-        CapabilityGrantId::new("grant.retained.session-refresh")
-            .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-        1,
-        grant_digest,
-        context.actor().clone(),
-        context.grant().issued_at,
-        expires_at,
-        resolved_scope.clone(),
-        BTreeSet::from([operation.capability_id().clone()]),
-        BTreeSet::from([operation.use_case_id().clone()]),
-        DisclosureClass::Evidence,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let lower_context = RequestContext::new(
-        context.actor().clone(),
-        resolved_scope,
-        grant,
-        context.request_id().clone(),
-        Deadline::new(expires_at).map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-        CancellationContext::active(
-            cancellation
-                .application_token_id()
-                .ok_or(RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-        )
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let binding = SessionRequestBinding::new(
-        identity,
-        capability_digest,
-        policy_digest,
-        configuration_digest,
-        cancellation,
-        budgets,
-    );
-    Ok(SessionRefreshCommand {
-        action,
-        context: lower_context,
-        binding,
-        target,
-        handle: selectors.handle.clone(),
-    })
-}
-
-fn admitted_action(
-    action: SessionRefreshActionV1,
-    handle: Option<&str>,
-) -> Result<SessionRefreshAction, RetainedSurfaceExecutionErrorV1> {
-    if handle.is_some_and(|handle| handle.is_empty() || handle != handle.trim()) {
-        return Err(RetainedSurfaceExecutionErrorV1::InvalidRequest);
-    }
-    match (action, handle) {
-        (SessionRefreshActionV1::Begin, None) => Ok(SessionRefreshAction::Begin),
-        (SessionRefreshActionV1::Status, Some(_)) => Ok(SessionRefreshAction::Status),
-        (SessionRefreshActionV1::Cancel, Some(_)) => Ok(SessionRefreshAction::Cancel),
-        _ => Err(RetainedSurfaceExecutionErrorV1::InvalidRequest),
-    }
-}
-
-/// Project-scoped requests must name the exact mounted project route. A
-/// profile-scoped request names no project; its binding to the admitted
-/// profile-session scope is checked through the resolved identity instead.
-fn request_matches_mounted_project_scope(
-    request: &SessionRefreshRequestV1,
-    context: &RequestContext,
-) -> bool {
-    let SessionRefreshScopeV1::Project { project } = &request.request.scope else {
-        return true;
-    };
-    let scope = context.scope();
-    let branch_matches = scope
-        .reference
-        .as_ref()
-        .and_then(|reference| reference.as_str().strip_prefix("refs/heads/"))
-        .is_some_and(|branch| branch == project.branch_id);
-    project.id == scope.project_id.as_str()
-        && project.repository_id == scope.repository_id.as_str()
-        && project.worktree_id == scope.worktree_id.as_str()
-        && branch_matches
-}
-
-fn admitted_identity(
-    request: &SessionRefreshActionRequestV1,
-) -> Result<ResolvedSessionIdentity, RetainedSurfaceExecutionErrorV1> {
-    let profile_id = ProfileId::new(request.scope.profile_id().to_owned())
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let store_id = SessionStoreId::new(request.session.store_id.clone())
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    let root_id = SessionRootId::new(request.session.root_id.clone())
-        .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    match &request.scope {
-        SessionRefreshScopeV1::Profile { .. } => Ok(ResolvedSessionIdentity::for_profile(
-            profile_id, store_id, root_id,
-        )),
-        SessionRefreshScopeV1::Project { project } => Ok(ResolvedSessionIdentity::for_project(
-            profile_id,
-            ProjectId::new(project.id.clone())
-                .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-            store_id,
-            root_id,
-            ResolvedGitRoute::new(
-                RepositoryId::new(project.repository_id.clone())
-                    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-                WorktreeId::new(project.worktree_id.clone())
-                    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-                BranchId::new(project.branch_id.clone())
-                    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-            ),
-        )),
-    }
-}
-
-fn admitted_target(
-    request: &SessionRefreshActionRequestV1,
-) -> Result<SessionRefreshTarget, RetainedSurfaceExecutionErrorV1> {
-    let temporal_mode = match request.target.temporal_mode {
-        SessionRefreshTemporalModeV1::Current => TemporalModeV1::Current,
-        SessionRefreshTemporalModeV1::AsOf { cutoff } => TemporalModeV1::AsOf {
-            cutoff: UtcMicros(
-                i64::try_from(cutoff)
-                    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-            ),
-        },
-        SessionRefreshTemporalModeV1::Evolution => TemporalModeV1::Evolution,
-        SessionRefreshTemporalModeV1::Forensic => TemporalModeV1::Forensic,
-    };
-    let grain = match request.target.grain {
-        SessionRefreshGrainV1::Occurrence => RetrievalGrainV1::Occurrence,
-        SessionRefreshGrainV1::LogicalMessage => RetrievalGrainV1::LogicalMessage,
-        SessionRefreshGrainV1::Turn => RetrievalGrainV1::Turn,
-        SessionRefreshGrainV1::Session => RetrievalGrainV1::Session,
-        SessionRefreshGrainV1::Thread => RetrievalGrainV1::Thread,
-        SessionRefreshGrainV1::Agent => RetrievalGrainV1::Agent,
-        SessionRefreshGrainV1::Summary => RetrievalGrainV1::Summary,
-    };
-    let frontier = SessionRefreshFrontierV1::new(
-        request.target.frontier.observed_through,
-        request.target.frontier.committed_through,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?;
-    SessionRefreshTarget::new(
-        SessionId::new(request.session.id.clone())
-            .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)?,
-        Some(request.source.scope.clone()),
-        temporal_mode,
-        grain,
-        frontier,
-    )
-    .map_err(|_| RetainedSurfaceExecutionErrorV1::InvalidRequest)
-}
-
-fn admitted_digest(domain: &[u8], material: &[u8]) -> [u8; 32] {
-    let mut digest = Sha256::new();
-    digest.update(domain);
-    digest.update(material);
-    digest.finalize().into()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{SessionRefreshAction, SessionRefreshActionV1, admitted_action};
-
-    #[test]
-    fn begin_requires_no_handle() {
-        assert_eq!(
-            admitted_action(SessionRefreshActionV1::Begin, None),
-            Ok(SessionRefreshAction::Begin)
-        );
-        assert!(admitted_action(SessionRefreshActionV1::Begin, Some("srh_token")).is_err());
-    }
-
-    #[test]
-    fn status_and_cancel_require_nonempty_handles() {
-        assert_eq!(
-            admitted_action(SessionRefreshActionV1::Status, Some("srh_status")),
-            Ok(SessionRefreshAction::Status)
-        );
-        assert_eq!(
-            admitted_action(SessionRefreshActionV1::Cancel, Some("srh_cancel")),
-            Ok(SessionRefreshAction::Cancel)
-        );
-        for action in [
-            SessionRefreshActionV1::Status,
-            SessionRefreshActionV1::Cancel,
-        ] {
-            assert!(admitted_action(action, None).is_err());
-            assert!(admitted_action(action, Some("  ")).is_err());
-            assert!(admitted_action(action, Some(" srh_handle")).is_err());
-        }
-    }
-}
