+++ crates/tracedecay/src/mcp/tools/handlers/graph/context_support.rs
@@ -309,13 +309,14 @@ async fn context_memory_matches(
         .search_project_memory_facts(query, read_control)
         .await
         .map_err(memory_application_error)?;
-    let mapped = crate::daemon::retained_owner::search_page(&page).map_err(|error| {
-        let problem = retained_surface_execution_problem(error);
-        TraceDecayError::Database {
-            operation: "project canonical context memory".to_string(),
-            message: problem.canonical_code().to_string(),
-        }
-    })?;
+    let mapped =
+        tracedecay_session_memory::memory_mapping::search_page(&page).map_err(|error| {
+            let problem = retained_surface_execution_problem(error);
+            TraceDecayError::Database {
+                operation: "project canonical context memory".to_string(),
+                message: problem.canonical_code().to_string(),
+            }
+        })?;
     Ok(ContextMemoryMatches {
         hits: mapped.hits,
         graph_coverage: mapped.graph_coverage,
